+++ src/pages/HomePage/Box/BoxCategory/StatusDefaultBoxCategory/StatusDefaultBoxCategory.tsx
