+++ src/pages/HomePage/Box/index.ts
@@ -1,0 +1,2 @@
+export { default as BoxCategory } from './BoxCategory/BoxCategory';
+export { default as BoxTodayTodo } from './BoxTodayTodo/BoxTodayTodo';
