+++ src/pages/HomePage/ModalContents/ModalContentsAlert/ModalContentsAlert.tsx
@@ -1,0 +1,11 @@
+import Complete from '../../../ModalContentsAlert/Complete/Complete';
+import DeleteAccount from '../../../ModalContentsAlert/DeleteAccount/DeleteAccount';
+import Logout from '../../../ModalContentsAlert/Logoout/Logout';
+
+const ModalContentsAlert = {
+	Logout: Logout,
+	DeleteAccount: DeleteAccount,
+	Complete: Complete,
+};
+
+export default ModalContentsAlert;
