+++ src/pages/HomePage/ModalContents/ModalContentsAlert/types/index.ts
