+++ src/pages/HomePage/ModalContents/ModalContentsFriends/FriendRequest/FriendsRequest.tsx
@@ -1,8 +1,8 @@
 import { userFriendData } from '@/shared/mocks/userFriendData';
 
-import ButtonRequestAction from './components/ButtonRequestAction';
-import FriendsListRequested from './components/FriendsListRequested';
-import InputSendRequest from './components/InputSendRequest';
+import ButtonRequestAction from '../../ModalContents/Friends/components/FriendsRequest/components/ButtonRequestAction';
+import FriendsListRequested from '../../ModalContents/Friends/components/FriendsRequest/components/FriendsListRequested';
+import InputSendRequest from '../../ModalContents/Friends/components/FriendsRequest/components/InputSendRequest';
 
 const FriendsRequest = () => {
 	return (
@@ -11,7 +11,7 @@ const FriendsRequest = () => {
 				<InputSendRequest />
 			</div>
 			<div className="h-[52.5rem] overflow-scroll">
-				<h2 className="subhead-bold-22 px-[1rem] py-[2rem] text-white">받은 요청</h2>
+				<h2 className="px-[1rem] py-[2rem] text-white subhead-bold-22">받은 요청</h2>
 				<FriendsListRequested friendsData={userFriendData}>
 					<div className="flex w-full items-center justify-end gap-[1rem]">
 						<ButtonRequestAction variant="positive">수락</ButtonRequestAction>
@@ -19,7 +19,7 @@ const FriendsRequest = () => {
 					</div>
 				</FriendsListRequested>
 
-				<h2 className="subhead-bold-22 mt-[4rem] px-[1rem] py-[2rem] text-white">보낸 요청</h2>
+				<h2 className="mt-[4rem] px-[1rem] py-[2rem] text-white subhead-bold-22">보낸 요청</h2>
 				<FriendsListRequested friendsData={userFriendData}>
 					<div className="flex w-full items-center justify-end">
 						<ButtonRequestAction variant="negative">요청 취소</ButtonRequestAction>
