+++ src/pages/HomePage/ModalContents/ModalContentsFriends/FriendsList/components/FriendInfo.tsx
