+++ src/pages/HomePage/ModalContents/index.ts
@@ -1,2 +1,3 @@
-export { default as ModalContentsFriends } from './Friends/ModalContentsFriends';
-export { default as ModalContentsCategory } from './Category/ModalContentsCategory';
+export { default as ModalContentsAlert } from './ModalContentsAlert/ModalContentsAlert';
+export { default as ModalContentsCategory } from './ModalContentsCategory/ModalContentsCategory';
+export { default as ModalContentsFriends } from './ModalContentsFriends/ModalContentsFriends';
